requirements.txt
boto3
awscli
Pillow
pypdf
langchain
streamlit
faiss-cpu
langchain_community